cleanup RuntimeDecorator
